requirements.txt
channels
channels_redis
django
django_mysql
djangorestframework
djangorestframework-jwt
django-cors-headers
pendulum
stripe
Pillow
